app/requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
attrdict==2.0.1
attrs==19.3.0
base58==2.0.0
bitarray==2.6.0
certifi==2019.9.11
chardet==3.0.4
charset-normalizer==2.1.1
customtkinter==4.6.3
cytoolz==0.10.1
darkdetect==0.7.1
docopt==0.6.2
eth-abi==2.1.1
eth-account==0.5.9
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.1.2
eth-typing==2.2.1
eth-utils==1.10.0
frozenlist==1.3.1
hexbytes==0.2.0
idna==2.8
ipfshttpclient==0.8.0a2
jsonschema==3.2.0
lru-dict==1.1.6
multiaddr==0.0.9
multidict==6.0.2
Naked==0.1.31
netaddr==0.7.19
parsimonious==0.8.1
Pillow==9.2.0
protobuf==3.19.5
pycryptodome==3.9.1
pypiwin32==223
pypref==3.3.0
pyrsistent==0.16.0
python-dotenv==0.21.0
pywin32==227
PyYAML==5.1.2
requests==2.22.0
rlp==1.2.0
shellescape==3.4.1
six==1.14.0
toolz==0.10.0
urllib3==1.25.6
varint==1.0.2
web3==5.31.1
websockets==9.1
xlrd==1.2.0
yarl==1.8.1
reportlab==3.6.12

